+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+  - Added support for index path components.
 
 ## [1.1.0] - 2020-05-31
   - Added `TinyTemplate::set_default_formatter` which, for example, allows to dissable HTML-scaping